+++ src/main/java/ru/yandex/practicum/filmorate/dal/interfaces/UserStorage.java
@@ -16,5 +16,6 @@ public interface UserStorage {
     long addUser(User user);
 
     User updateUser(User newUser);
-    void deleteUser(Integer userId);
+
+    void deleteUser(long userId);
 }
